# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Documentation for Warehouse Inventory Management System",
        "title": "Warehouse Inventory API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/pembelian": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of all purchase transactions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pembelian"
                ],
                "summary": "Get all purchases",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PembelianResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.ValidationError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new purchase transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pembelian"
                ],
                "summary": "Create new purchase",
                "parameters": [
                    {
                        "description": "Purchase Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.BeliHeaderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.PembelianResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ValidationError"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/middleware.ValidationError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.ValidationError"
                        }
                    }
                }
            }
        },
        "/api/pembelian/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get details of a specific purchase transaction",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Pembelian"
                ],
                "summary": "Get purchase by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Purchase ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PembelianResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/penjualan": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a list of all sale transactions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Penjualan"
                ],
                "summary": "Get all sales",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PenjualanResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new sale transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Penjualan"
                ],
                "summary": "Create new sale",
                "parameters": [
                    {
                        "description": "Sale Request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.JualHeaderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.PenjualanResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/penjualan/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get details of a specific sale transaction",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Penjualan"
                ],
                "summary": "Get sale by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Sale ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.PenjualanResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/middleware.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "middleware.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "middleware.ValidationError": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.BarangPembelianResponse": {
            "type": "object",
            "properties": {
                "kode_barang": {
                    "type": "string"
                },
                "nama_barang": {
                    "type": "string"
                },
                "satuan": {
                    "type": "string"
                }
            }
        },
        "models.BarangPenjualanResponse": {
            "type": "object",
            "properties": {
                "kode_barang": {
                    "type": "string"
                },
                "nama_barang": {
                    "type": "string"
                },
                "satuan": {
                    "type": "string"
                }
            }
        },
...
        "models.BeliDetailRequest": {
            "type": "object",
            "properties": {
                "barang_id": {
                    "type": "integer"
                },
                "harga": {
                    "type": "number"
                },
                "qty": {
                    "type": "integer"
                }
            }
        },
        "models.BeliDetailResponse": {
            "type": "object",
            "properties": {
                "barang": {
                    "$ref": "#/definitions/models.BarangPembelianResponse"
                },
                "barang_id": {
                    "type": "integer"
                },
                "harga": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "qty": {
                    "type": "integer"
                },
                "subtotal": {
                    "type": "number"
                }
            }
        },
        "models.BeliHeaderRequest": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BeliDetailRequest"
                    }
                },
                "supplier": {
                    "type": "string"
                }
            }
        },
        "models.BeliHeaderResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "no_faktur": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "supplier": {
                    "type": "string"
                },
                "total": {
                    "type": "number"
                },
                "user": {
                    "$ref": "#/definitions/models.UserSimpleResponse"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.JualDetailRequest": {
            "type": "object",
            "properties": {
                "barang_id": {
                    "type": "integer"
                },
                "harga": {
                    "type": "number"
                },
                "qty": {
                    "type": "integer"
                }
            }
        },
        "models.JualDetailResponse": {
            "type": "object",
            "properties": {
                "barang": {
                    "$ref": "#/definitions/models.BarangPenjualanResponse"
                },
                "barang_id": {
                    "type": "integer"
                },
                "harga": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "qty": {
                    "type": "integer"
                },
                "subtotal": {
                    "type": "number"
                }
            }
        },
        "models.JualHeaderRequest": {
            "type": "object",
            "properties": {
                "customer": {
                    "type": "string"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.JualDetailRequest"
                    }
                }
            }
        },
        "models.JualHeaderResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "customer": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "no_faktur": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "total": {
                    "type": "number"
                },
                "user": {
                    "$ref": "#/definitions/models.UserSimpleResponse"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.PembelianResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.BeliDetailResponse"
                    }
                },
                "header": {
                    "$ref": "#/definitions/models.BeliHeaderResponse"
                }
            }
        },
        "models.PenjualanResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.JualDetailResponse"
                    }
                },
                "header": {
                    "$ref": "#/definitions/models.JualHeaderResponse"
                }
            }
        },
...
        "models.UserSimpleResponse": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}